Create Linear Regression
By using data from Forbes I:
Fit·a linear regression model to these data using profits as the dependent variable and sales and assets as the independent variables.
Analyzed the residuals to check the adequacy of the model. Computed the leverages associated with the data points. 
Generated a 95% prediction interval for profits corresponding to sales of 100 (billions of dollars) and assets of 500 (billions of dollars).
Carried out a likelihood ratio test of Ho: {32 = 0 with a significance level of a = .05.