Create Bones
Performed a regression analysis by fitting the response for the dominant radius bone to the measurements on the last four bones.
Suggested and fitted appropriate linear regression models. Analyzed the residuals.
Performed a multivariate multiple regression analysis by fitting the responses from both radius bones.
Calculated the AIC for the model and for the full model.